job validation complete
