+++ src/routes/user_routes.js
@@ -2,7 +2,8 @@ import { Router } from "express"
 import { UserModel } from "../db.js"
 import bcrypt from 'bcrypt'
 import e_auth from '../middleware/e_auth.js'
-import { check, validationResult } from 'express-validator';
+import {validationResult } from 'express-validator';
+import {newUserValidate} from "../middleware/validations.js";
 
 
 
@@ -12,34 +13,6 @@ const router = Router()
 router.get('/',e_auth, async (req, res) => {
     res.send(await UserModel.find({}))
 })
-
-// Validation for user creation
-const newUserValidate = [
-    check("email", "Please use a valid email address")
-        .trim()
-        .escape()
-        .isEmail()
-        .normalizeEmail()
-        .custom(async(value) =>{ // Ensures no dplicate emails
-            const existingEmail = await UserModel.findOne({email: value})
-            if (existingEmail){
-                throw new Error("Email already exists , please check Employee list")
-            }
-        }),
-    check("password", "Please ensure password is at least 10 characters and has a number")
-        // .isLength(10)
-        .trim() // removes white space 
-        .escape(), // Changes HTML symbols for security reasons
-        // .matches(/\d/) // macthes a number
-    check("name", "Name must be at least 3 characters and first and last name sepearted by a space")
-        .isLength({min: 3})
-        .trim()
-        .escape()
-        // .custom(async (value) => {                        // ensures first and last name (at least two words)
-        //     const wordInName = value.trim().split(/\s+/)
-        //     return wordInName.length >= 2
-        // }),
-]
 
 
 // create user (Admin only)
